Try to add support for dedup.
The file flip2.gfa shows why this is needed.
Basically we don't want to add the same link twice
even if _two_ new paths require the same link, and,
generalizing, we don't want to add an existing link to the graph
just because a new path needs it.

I'm running into weird type errors though...