Go back to full range of tests: fetched and handmade
We agree with odgi on all the handmade files except note5.gfa
I think odgi is doing something fishy there;
see https://github.com/pangenome/odgi/pull/485#issuecomment-1496693803

We agree with odgi on all the handmade files except flip4.gfa
flip4 is not a valid graph and so arguably it does not matter what we do with it,
but `odgi flip` does currently _fix_ flip4, so I think it's worth keeping around.
The point of divergence is that we do not currently fix graphs like flip4, though it's easy enough to do if that were desired.
See https://github.com/pangenome/odgi/issues/496