Use Turnt's new `out_dir` for validate tests
The problem with these tests is that they needed to use files with a
`.gfa` extension, but that would create files that we would then treat
as tests! So I added a feature to Turnt for sequestering these files in
a subdirectory so they don't get confused for tests:
https://github.com/cucapra/turnt/pull/27

This change uses that feature to make these "invalidated" graphs more
like normal tests.