Always run all slow_odgi tests
Currently, the "main" tests fail because of the whole "flip" thing. As
indicated here:
https://github.com/cucapra/pollen/issues/119#issuecomment-1635042849

This stops the rest of the test commands from running. Now we run
everything, at the cost of telling Make to ignore failures. This is
probably the right thing to do in general so we don't hide possible test
failures that happen after the first one.

This is another problem that would be solved if this could be one big
`turnt` invocation.