Typed Id & Span
An experiment to see what it buys us to have the Id and Span types carry
with them the type they are supposed to reference. So you can think of
`Id<Segment>` as kinda like an arena-fied version of `&Segment`, and
`Span<Segment>` as like an arena-fied `&[Segment]`.

This in no way helps you avoid using an Id with a different arena of the
same type. But at least you can no longer confuse arenas of different
types.