Make PackedSeqStore::create borrow instead of own
Previously, `PackedSeqStore::create` took *ownership* of a
`Vec<Nucleotide>`. In Rust, this means that it *consumes* this vector as
input, destroying it while building the compressed data store.

For this `create` method, there is no need to consume the input, since
we are creating a copy of all the data anyway (in order to compress it).
So it makes more sense to borrow the input, as opposed to taking
ownership of it.

So I could have changed this to take a `&Vec<Nucleotide>` as an
argument. But in general, when the code doesn't actually care that the
input is a `Vec` but just needs some sort of input array, it makes more
sense to take a slice, `&[Nucleotide]` here, instead: because a
`&Vec<T>` automatically converts to a `&[T]`, it's effectively a "more
general type," allowing this method to be used in more contexts. So
that's what this method now does.

The eventual upshot is that the `test_export_import` test does not need
a `clone` to copy the entire vector in order to make the assertion at
the end work.