packedseq: Display like a string, not a list
This changes the `Display` impl for `PackedSeqView` to emit dense
strings of nucleotide symbols like "ACTGG" instead of comma-separated
lists like "[A, C, T, G, G]". I contend that this is what humans
actually want to see if they happen to print out a nucleotide sequence,
and it helps with testing too (we can make compression/decompression of
sequence data a byte-for-byte round trip).