packedseq: Avoid an intermediate vector allocation
When compressing ("exporting") a sequence, we previously allocated an
intermediate vector for `Nucleotide` values and then fed that into
`PackedSeqView::create(...)`. We can avoid this intermediate value by
directly iterating over the characters in the input. This seems to come
with a modest performance win.

To make this work, I introduced a new `from_iter` constructor, renaming
the old constructor to `from_slice` to differentiate it.