packedseq: Treat input as ASCII
Previously, when parsing the text for an input nucleotide sequence, we
were parsing it as a string, i.e., Unicode text encoded as UTF-8. We in
fact only need individual bytes, so I have introduced a separate path
for interacting with plain ASCII bytes. This has come with a roughly 20%
performance improvement when exporting in my limited testing.